Day 0 cleanup
This is a small simple commit whose purpose is to mark the start of the
project after doing cleanup and getting ready for things.

Just to start us from an even keel, this inserts Svelte 5 into the mix
so that we can generate components as desired later and know that all
of the machinery is set up, since this does not seem like a fun way to
spend the first day of devember.