Start the abstraction of the command interface
The sample plugin started with a large plugin that includes a couple of
commands whose bodies are inlined directly into the body of the plugin.

This is good for illustration purposes, but it's a maintenance nightmare
since it makes the plugin file unwieldy (probably).

This is the start of a simple abstraction to pull these out of the
plugin body itself.