Enhance the command factory abstraction
This enhances the logic already created for command factories to provide
the ability to have a simple list of commands so that the plugin can
use a short loop to include them all rather than a large number of
calls to the add handler.

In addition this also includes the notion of some combined check
function factories that allow for and(), or() and not() functionality
in checks, as well as a minimal number of check functions to get us
rolling.