Start refactoring for allowing Editor commands
Obsidian has two types of commands; one of them is what we would refer
to as a "standard" command, which can execute at any time of your
choosing, and an "editor" command, which examines and can modify the
state of an editor, such as grabbing text or making changes.

There is more to do here since this does not yet cover the notion of our
combiner functions, which need to be made more generic, but it is a
good start for allowing editor commands regardless.