Allow for generic combinators on command checks
The previous implementation of command checks assumed that all checks
get the same arguments no matter what (a plugin). For editor commands,
we actually get the editor and the view to work within as well.

This is fine except that for our combinator functions and(), or() and
not(), those assume always a single argument.

This command makes those functions take arbitrary arguments and will
just pass them along.