Include a simple implementation of a custom view
This commit sketches in the code required to register a new view type
and tie the ribbon button to opening it.

This is a minimal possible example and probably needs more juice to it,
such as controlling whether or not the view can be split or whether it
can be put into the normal content area.