Implement a version of state savings in our view
This could do with a bit of cleanup and creating appropriate types for
everything, but this is the MVP version of having a custom view store
data that it uses in its display into the workspace file in Obsidian
and have that restored back when the workspace loads.

What this does NOT cover is saving data that is distinct from this,
since this data is ephemeral to the view and will go away when the view
is closed.

Hence if we need to persist any data that the user is interested in,
that needs to be loaded from a different location.