Fix LSP issues plaguing the new code
In a turn of events that makes me wonder how or why anyone uses either
LSP or Typescript in general, this resolves the issue whereby the code
builds just fine when using a type annotated call to mount() even as
that same code causes LSP to crap right in its little electronic
underpants.

The reasoning for this is apparently because:
 1. esbuild largely ignores the tsconfig.json, or otherwise does not
    care if it generates type errors (which is midly distubring)
 2. LSP cares about tsconfig.json really a whole lot, and gets very
    angry.

In a perfect world, a configuration file which is used by the compiler
of the language would be used by **ALL** of the tooling in the project
that uses said language, but apparently the world we live in flawed.

Or people in charge of tooling have their heads solely right up inside
of their own asses for some reason.