Start the construction of a Settings page builder
Rather than a large pile of nested, obscure function calls, this is a
first crack at a factory that can create the appropriate settings page
from a data object.

The theory is that this makes the settings configuration easier to
setup and maintain because all of the boilerplate is abstracted away and
the configuration becomes largely data driven, while still allowing for
more custom solutions.