Enhance the setting layout factory code with types
The weirdness of that statement notwithstanding, this changes the
factory that we previously implemented, but pushes the responsibility
of the settings get and set from the object upwards from the config
into the factory itself by way of passing in a settings manager
object (the plugin itself).

This keeps the config small and removes the need for it to have access
to the plugin itself, which means that it could in theory be loaded
from a file or exist outside of the code of the class that needs it.

This was a little hard fought given that even though I wrote a note of
how to implement this, when I actually did I ignored part of it and
spun my wheels for a bit.