Clarify our setting headers
The API docs mention a Header settings type, but using this does not
actually create anything that looks like a header. In particular it
counts as a single setting, which gives a border in places we don't
want, and it is not larger than anything else.

This small tweak to our factory adjusts the CSS we ship to remove the
top border from such an item, as well as make its font bigger.

Also, in order to make our lives easier since we have set things up to
assume that config items of type 'header' are headers, the CSS class is
applied to them automagically if no other CSS class is assigned.