Consolidate our data loader to hold other data too
The loadData method in plugin is actually loading the plugin's data.json
file; while many plugins use this as an overall data dump, my OCD very
strictly requires that the settings be in a subkey named settings so
that they are distinct from other data.

This change implements that, and changes some interface names so that
they make more sense.