Fix annoying build issue
Starting at the point in time in which we altered things to use a type
annotated call to mount() to mount our components, the build command was
inherently broken because the typescript compiler seems to ignore the
plugins within the config that tell it how to deal with svelte.

So where we had at one point a problem in which the code was correct but
the LSP did not think so, we also had a problem in which the code wad
correct but the production build did not think so.

The dev build was unaffected because of course esbuild does not bother
to look at the typescript config at all and seems more or less blithely
unaware of issues on the whole.

The fix is to not use tsc to do the check and instead use svelte-check
instead. For this reason, new tool scripts were added.

This change was made outside of stream in order to facilitate the
cleanup commits that follow.