Clean up imports
I went through all of the imports in all of the files and put them into
some form of order; external things go first, then we go second. Also
types go first, other things last.

We inline the type in an import (but they come first in the import) to
try and minimize importing from the same file twice, and pull the type
to outside the braces if the whole import is types, for clarity.

This was not done on stream since it is just busywork.