Rename our model file to types and add hydration
Just because in the cold light of a new day the word model has
connotations of MVC architecture and what is in that file is not
technically that (at the moment) the file has been renamed to better
track what it contains, pending another mental flip-flop.

This also includes some hydration code to fill out an empty object with
first defaults and then overlaying the loaded data on top, so that when
the function is called, all values have a sensible default and will be
present.