Reorganize our factory code into a new directory
Rather than having the two factory systems in the utlity code with
module names that bely their functionality, a directory that represents
their functionality was created and they were moved over and renamed.

This makes things cleaner since it is more clear that you want to look
at a factory folder when wanting to look at or work with factories.