Refactor the commands for a better file structure
This change alters the layout of the commands so that they follow a file
structure that's more logical and easier to find what you're looking
for.

The main thing is that the checks and combiners are part of the command
structure (although arguably the combiners can be used with any kind of
function that returns a boolean), so they should live there.

Similarly since we have two different kinds of commands, those should be
split out as well so that it's easier to find them and tell what they
do just from their names.