Rename the sample component to a sample modal
We are about to implement a second component similar to the first but
for use directly in the view, while we will leave the modal content the
same as it is for the time being.

As such, the names of things needed to change first to ensure that
things still work as expected.