Implement a somewhat flawed runic update system
This change is a modification of what we previously had for the view
that stores some of its data (i.e. the click count) into the workspace,
such that this also happens but also we can modify the other data in
our data.json file.

This is done via runes so that code that cares to know about the state
change will just be invoked automagically, rather than having to pass
us a function directly which we then call.

This is currently a tiny bit janky because it's triggering too many
updates (every button click to increment the count is needlessly
writing the data.json file) but it works, so we will leave it as is for
now.