Fix some linter related issues
There were some linter issues in the last commit made yesterday that
were not blockers, so we let them slide; this fixes them.

Two of them related to interfaces that were empty, which is not a
desirable thing in general but in this case it was intentional to our
pattern, so we left it. This is now disabled via magic comment.

The other was that by default eslint uses everything to the right of the
first dot in a filename as the extension, so anything that
is .svelte.ts is treated as JavaScript and not Typescript, because that
extension is unknown. This causes the linter to have issues because
such files have type annotations that it chokes on. This is fixed by
hackily telling the linter that such files are typescript.