Update our persistence layer to be more generic
This commit crudely extends the persistence layer that we created
yesterday to be more generic so that it splits the session persistent
data and the plugin persistent data into alternate state variables.

This ensures that updates of one don't unintentionaly invoke an update
in the other one too.

There is still a bit of a weirdness here because some of the debug logs
show that the state they are passed is not the new state but the old
one, yet things still work. This will take more puzzling to figure out.