Update our logging to work better with proxies
We were noting during yesterday's changes that while things seem to work
just fine, the logging of the incoming state values was always wrong
and seemed to reflect initial state and not the full state. This was
not the case when the implementation was not generic, and there the
state was tracking specific variables instead of an object.

On a hunch I altered the logging to dump a JSON string version and
the problem is resolved.

Best guess theory at the moment is that Svelte is being clever about
updates and is proxying individual values as well as the object, and it
knows it does not need to update the whole object for a single field or
similar, so the data that is logged is wrong until it is forced. Just a
guess though.

In any case, this seems to fix things so.