Make our state handlers skip the initial update
Based on how things are currently structured and interacting with the
Obsidian API, the effects that track the updates to plugin and session
data run immediately during initialization, which is suboptimal.

This little bit of probably unwise hackery causes the code in the first
effect run to be skipped since it's just providing the information that
is already available.