Alter the session save code to be more piecemeal
Previously our solution for allowing session and plugin data from a
svelte component to persist required that there was always session data
and always plugin data, even if you did not need them.

This change makes both optional so that it's possible for a component to
use one or the other and not just both or none.

This needs a wee bit of cleaning to be less ugly looking, but the
functionality is there. We still to decide how to best utilize this.