Implement an initial crack at a view base class
Using the integration we just implemented (which is as yet fully
untested), this implements a base of as view class for us to descent
from instead of ItemView so that we can easily mount a component and
have the the session and plugin data that is needed by that component
automagically update without any code.

The idea is that this will reduce significantly the boilerplate needed
to do this, so that if we have many views, we are not repeating the
same code everywhere.

This is as yet not rolled into anything, but this will happen tomorrow.