Update our view base class handlers
Without actually doing any testing, this extends our previous base class
for Obsidian view data persistence to also onboard the methods required
to persist the session data for us.

This leaves us with only having to worry about saving plugin data,
although if we changed the plugin type from generic Plugin to our
plugin or enforced that it had a specific method in it, we could get
away from that as well.