Apply the new base class to the sample view
This change, which for a wonder seems to work, updates our sample view
to use our new base view class.

This keeps it fully functional without having all of the annoying
boilerplate code in there; instead we just need to return the
component, its properties, and what the initial data should be.

That could also be tightened up most likely, but we will leave it be for
now.