Use our newfound powers to make dynamic status bar
As a test of the system, this small commit creates and uses a Svelte
component to represent the status bar and easily update the currently
focused tab display in it.

One could imagine that the status bar could be more reactive, such as
displaying different things depending on the type or location of the
file and such, without having to have that logic directly within the
plugin code itself.

This also gives the status bar the theoretical ability to both read and
update the saved plugin data as needed, and to persist any data that it
deems interesting into the session for later.