Update the Svelte integration mount logic
This is a small in place change but it modifies the mount() signature in
the integration class to have multiple variations, one with full
arguments and one with an options argument.

This allows for the constructor to have an easier call; we will be
making one further cleanup here, but it's scary so we're gonna do it
separate.