Include a sample of a custom code block handler
This needs some work because things are not as reactive as we might like
using the current abstraction, but this showcases a minimal viable
solution to having a svelte component be mapped to process the input of
a code block.

This uses an additional handler class, which would be the place for
exposing any functionality related to parsing the content of the data.

More work will be done on this tomorrow.