Refactor the custom markdown child test code
Small change but since the stuff we have mostly kind of works if you
don't count reactivity, pull that out into its own file so that it is
not inlined within the plugin code directly.

This also modifies said code so that the code block keeps track of what
language it was created for, in case that is useful later.