Create a code block factory setup
Similar to our other factories, this sets up one that can easily make an
association between a code block language and the class that is
intended to render it.

This makes the plugin code cleaner. As a side note, this is a slightly
cleaner factory than the function factory is, so perhaps we should
adjust that one too.