Small tweak to the command factory
To make the command factory look and behave more like the factory we
made for code blocks, alter the pattern slightly. For blocks we want to
have the registration happen right inside of the handler since it is a
weighty piece of code, so might as well do that for the commands as
well.

This also cleans up a couple of linter warnings from the last commit
last night, where I forgot to double check before I signed off.