Add the notion of Ephemeral mount data
Just to round out the things that may be interesting to us here, this
adds the notion of purely ephemeral data; fields that a view may want
to bind to a svelte component that need to be reactive between the two,
but for which there is no need for saving that data anywhere in the
session or in the plugin data.

This is surfaced via a toggle button in the sample panel view which,
unlike the other two controls, never persists anywhere.