Make the text area style properly
I could swear that I tried this previously and it did not work, although
I did not use the box-resizing thing; can't imagine how that matters,
though.

I highly suspect I got caught up in thinking that the CSS would be
dynamic when the plugin reloaded, which does not seem to be the case.

In any case, this also makes the button have an appropriate button style
as a primary button and, since saving does nothing when the data has not
changed, also leverages reactivity to ensure that the button is disabled
when it won't do anything.