Alter our data structures for svelte data shares
Previously this required three distinct interfaces be provided, with
undefined being set in the generic type parameter for any that are not
required.

Now there is a single schema item which carries all of them at once and
you only need to define the ones you use, which makes everything a lot
smaller.