Do a similar thing with the component properties
Similar to the previous commit, this uses a single component schema type
object to convey both the properties and the exports of any Svelte
components, so that the number of types you need to pass around is not
quite so insane.

This actually caused a bit of an issue with the Svelte code because in
order to get the types of its props, it seemed a bit hairy. However I
found an example that showed a helper for that, so we;re rolling with
that now.

It is an extraneous export, but at least the component only needs to
import a single type to know what it's doing, so there is that.