Set up a global plugin data handler for more juice
In particular, this commit achieves the goal of having a single source
of truth for the plugin data that may or may not be handed off to
Svelte components that want to track plugin state.

By doing this, any component that uses it will see changes when any
other component makes a change to the data; this allows things like our
code block to update itself whenever the content data changes without
any extra code.

The plugin now sets up a watcher for changes to the data structure and
responds by saving, so there is no need to do this from within the
sample view Svelte component any longer.

This leaves a bit of a hole if there was ever a case where the plugin
data should be updated but, for whatever reason, not persisted. However
that seems like a case for shared ephemeral state, so.