Make the sample modal use new ephemeral state
This updates the sample modal so that it uses the svelte integration
handler. Additionally the component now exercises the ephemeral data to
keep track of the number of times the button was clicked during the
session.

This also showcases that we should have a base modal class the same as
we have a base ItemView class so that things could be more
parameterized and less gross.