Implement a base Modal class that supports Svelte
Similar (very similar, since it's a copy/paste more or less) to the
ItemView Svelte base class we have that allows the creation of an item
view that mounts a Svelte component without a lot of boilerplate, we
now have one for modals as well.

The sample modal was also updated to use it.

Note that our ComponentSchema seems to require the props and exports
even if we don't need them, so we should probably look into that and
see if there is a fix.