Implement the handler for data.json updates
It turns out that there is an optional method in the Plugin class that
will get invoked whenever the data.json file is updated externally.

We now implement this and reload the data; this will also update the
shared state for plugin data that all components may bind to, which
will cause them to get the update right away as well.

Note that for this to work the plugin folder needs to match its ID in
the manifest file; this is always the case when installed but may not
be when you're doing local development.