Bump our minimum version to the latest one
This is technically the most recent Catalyst version, which means that
the public in general does not have access to it yet. However it has
functionality that we want to target, so we're bumping ourselves now.

As a part of this we have also updated the API wrapper we have to the
latest version, although since this build is no public yet there is no
documentation on what has changed.