Update our settings to use a SettingGroup
As of version 1.11, the version we just jumped to, the API supports a
new thing called SettingGroup whose job it is to group settings; this
was previously the purview of the plugin author and one presumes that
it took the maintainers a bit to realize what a bad idea it is to leave
that up to the developer, for reasons of layout consistency.

This should theoretically lay things out as they worked previously, but
using a new layout (which no longer requires custom CSS, thank god,
cause that skeeved me out).