Augment the settings to include a toggle item
Toggle items are booleans, and with our current structure all set up, we
can now easily add them to our config "schema".

More things still need to be added in here, although we need to be maybe
smarter because it's possible to include more than one setting in each
setting item (such as buttons that do things) and we currently have no
way to deal with that.