Fix the ability to disable commands
Not sure if this was always broken or it was broken when we switched to
using SettingsGroup, but totally setting this on the actual setting
object, which is apparently allowed, has no effect whatsoever on the
actual controls.

On the other hand, setting it on the controls as they're added works,
but does not visually distinguish them.