Include some CSS to mark our controls as disabled
For any controls that we generate, ensure via CSS that when they are
disabled, they appear disabled; we make them grayscale and dim them out.

In theory this works for any color scheme and in light or dark, although
it has only been tested for light and dark using the default Obsidian
theme.

It remains a mystery why Obsidian allows you to mark things as disabled
but will not in fact make them look disabled.