Phase one settings factory code cleanup
In order to get the settings factory looking decidedly less gross since
as a single file it was only going to get more and more unweildy as
time went on, this is a phase one refactor that passes the
configuration items to specific handlers for different control types to
be added to the setting.

This not only cleans up the code but also gets us better prepared to
handle the upcoming situation in which we may want to have multiple
controls per setting.

There is more to do here since the code is not as clean as it could be
but this sets the baseline.