Include the TextArea setting type
This one is currently a little bit problematic because it defaults to
resizable and the resizing is all wonky. Probably best to style these
with some kind of CSS class that forces this to not be the case, but
for now that is low hanging fruit because one presumes that the default
should allow it to not make the settings page look ridiculous.