Allow the toggle to have a tooltip in the config
Previously we played with trying to get tooltips to work and none of
them did but as it transpires that's because we tested before we added
the toggle, and it appears to be the only component that actually
supports it.

We have added this such that the tooltip can be either a single static
string that indicates the tooltip to always use, or an array of strings
should you want the tooltip to change depending on the toggle (i.e. to
give clues about what the current setting actually means for the user).