Update how our TextArea works to stop the funk
When we added the TextArea to the settings types, what was discovered
was that the resize is unconstrained in both directions, but it resizes
not in the normal way, but as if it is resizing based on the center of
its container.

This constrains the resize to just be vertical for sanity while we work
this out.