Include a quick progressbar setting component
This component is meant for being able to display a percentage value;
the API for this on our end is still TBD because we need to have all of
the settings be able to adjust their values dynamically based on other
values changing.

What we have right now is a handler function that can be invoked in
order to collect the value to display. This is currently triggered only
one time; we can adjust later as needed.