Include a Slider setting to the factory
The slider is pretty bare bones in the API and probably needs to be
composed with other things in order to be more useful. Here we can set
the min/max/step values.

THe API uses 0/100/'any' as the defaults for things you do not specify
which gives you a percentage that has no constraint, so you can end up
with fractional numbers.