Include a colorpicker setting component
This is currently a little bare bones in as much as it assumes that the
color model to use is Hex "#rrggbb" format even though the component
also allows for RGB and HSL color types values, which are simple
objects that have fields.

For now, and for our purposes gathering stuff for the config file, the
RGB is enough, particularly because it requires fewer bits of nonsense
to get going.