Fix some wacky types issues
When following through on our types pattern from the initial code for it
to the other places, forgot to mark that the schema and component were
subclasses of the required types, which causes a couple of issues.

Also, since some things don't need exports, we can solve the problem of
having to declare the object as having them by just not using that part
of the type when we declare the getComponent() return type. Alternately
we could use the helper or pre-export the Exports like is done for
Props, but since Props is needed for the actual svelte component, this
way seems cleaner.