Include a little bit of dynamic settings setup
This is an attempt at a system by which one setting can declare that it
depends on other settings for its value, so that whenever any of those
settings change value, it gets a change to change how it's configured.

This is currently prototyped using just the dropdown control, which can
listen for any setting to change and rerun its setup. In our case the
state of the toggle changes the text of the options.