Allow for the progress bar to also have deps
Now that the scaffold is in place, this updates the progress bar so that
it can also have its value altered depending on the value of other
settings being changed.

In our example here this causes the progress bar to jump to the value of
the slider setting (once you let go of the thumb).

More than likely the progress bar is actually monitoring some external
process, but if one of the settings controls the process, knowing that
a setting has changed might need to result in a re-calculation, so.