Version 1 of multiple controls per setting row
This is a small type based change and a small structural change that
allows for more than a single setting control to appear inside of a
single settings row (although singles are still possible).

This is done in a way that is backwards compatible with the existing
setting config to make life easier, but we should adjust this to make
it better. That can happen after we add in a couple more controls so
that we have a better example of it.