Include button controls to the settings factory
Obsidian supports two different kinds of buttons; normal and extra. An
extra only has an icon, and has no button frame. A normal can have text
OR a button, and can be normal, warning, or call-to-action (cta)
styled.

Our examples have been augmented to include examples of both of them.