Version 3 of multiple controls per setting row
Small, small tweak to the previous. Use a discriminated type for the
distinction between the two row types; this way you can have an entry
that is a heading OR an entry that is a settings row, without them
being combined.

THis makes rearranging settings easier than if the first setting in the
group also holds the heading, because then you would need to do more
work.