Include the MomentJS Format control
This is a simple control that allows you to select a moment date format
and optionally visualize the result (which is why it's not just a
simple text field; it does a bit of heavy lifting for you).

The API is kind of annoying because the default format that you give is
also the placeholder, and the control uses that format for
visualization if the value is not set, which means that if you don't
provide a value your code needs to keep track of what the default was.

This is sub-optimal, so instead we assume that if the value is the
default value, then store that even if that's not the value.