Allow passing the plugin to things that need it
For the upcoming work on the Search component, we need to be able to
pass the app object through to a settings factory handler. If we need
that we should instead pass along a plugin instance instead, because
the plugin is also useful for things like buttons, and it holds a
reference to the application that created it as well, which is a
twofer.

As a part of this we have augmented the dropdown handlers to take this
as well as the button handlers. Buttons might want to invoke some
method on the plugin, or on a property of some element of the plugin,
and the dropdown might have complex requirements for gathering the
options that is best expressed not directly in the config.