Implement an initial search component
This is essentially a text component with an ability to search attached
to it. A class instance is required to actually pull this off, which
gets told the query string every time it changes, allowing it to filter
the list of items by the query.

Like text fields, this hammers the setting save every time you type in
the search field, but there may be no good way around that.

We should also probably parameterize this so that it is not forced to
just be simple strings; that way we have more control over what it's
inserting or what the source is.