Remove redundant type extends clauses
This is a small change but after committing and pushing last night I had
this thought (go figure) and did a little bit of research and realized
that I was using `P extends Plugin` in more places that are needed; in
the type definition,it's just a placeholder. The places where it's used
in a more concrete way (the factory, etc) are the only places that need
it.

So, this just makes it a little nicer to look at.