Quick stab at making search more expansive
This is a small tweak which works but probably not with the type safety
that is desired, to allow the search field to work with any sort of
field and not just strings.

Thus the examples are modified in a ham handed way to allow for both a
string and a number.