Modify the textarea setting to allow line/resize
The TextArea field now supports being configured with both a resize and
a lines value, to specify the number of lines in the control and also
whether it can resize horizontally, vertically, both or neither.

We toyed with the idea of being able to set the columns too, but that
can easily cause the control to jump outside of the bounds of the
settings modal without more tweakage, and I don't know that that
tweakage is worth it at the moment.