Implement a simple debounce on plugin data saves
This makes the call that is intended to trigger a save of plugin data
debounce so that frequent calls (such as from the onCHange handler of a
text field in the settings page) do not hammer the API with requests to
save. This still keeps the calls to trigger settings listeners so that
they happen right away, even before the save is complete, since
otherwise we would block dependent field updates.

I did some searching and this does not seem to be debounced internally,
despite their settings examples clearly saying to work things a certain
way.

So this is either a good idea, or I am a grumpy old man shaking his fist
at clouds right now.