Make the disabled state a function and not bool
Previously the disabled state was just a boolean, but it should actually
be a function. Partially this is because the state should not be hard
coded, as it would be that way, but also since we are working towards a
more complete system for doing updates on fields based on their
dependencies changing, this state may well change if it depends on a
sibling setting.

To that end, the static dropdown sample is now only enabled while the
toggle is enabled, to show that this has an actual effect on things.