Try a different take on search stuff
This took a lot of futzing around, but the gist is that the search can
now use arbitrary objects for the search values (not just strings or
numbers), which allows for the possibility of having a richer display,
at the cost of a type machination that makes me weep.

Currently, for the custom number example, the string that gets displayed
back into the setting box is the underlying ID value, because that is
what gets saved to the settings, and when opening the settings pane
again that's what you're gonna see. So this makes for better UX in the
general state, since it does not have to try to refetch the options just
to map the initial display.

That could maybe be an option in cases where there is not anything async
about the items being fetched.