Fix the previous commit's boneheaded blunder
Due to accidentally including the word const where I did not intend to,
I managed to break the whole point of the feature where it parses the
value and ensures that one of the proper valid values is present by
masking the outer variable that is otherwise used to track the parsed
output. Whoops.