Initial calendar svelte component
This commit introduces a new calendar code block that can display (in
theory) any calendar day for any year. I say in theory because
currently it wholly ignores the content of the code block and hard
codes in the actual calendar dates used.

However, for a wonder flex and grid were both used and for the first
time in many years, they actually worked without being a gigantic pain
in the ass.