Include simple parsing for the calendar component
There is still more work to do on the calendar, but for now this makes
it more functional by actually allowing the code block to specify the
calendar name and the dates involved.

In actual use in the plugin there would probably be some sort of course
and the calendar could focus and have controls. but this is good enough
for the time being to show intent.