Tighten up the calendar display and add a Today
This change does two things. First, the code that generates the cells to
display is greatly tightened by pre-allocating them all and then just
filling in the days; this saves needing three loops.

The second is the addition of some styling that is used so that if a
calendar happens to have an entry for today, it is marked as such.