Update the source parser in the Calendar
This very slightly enhances the parsing of the content of the calendar
code block so that it's more robust and easier to understand while
you're looking at the markdown.

For no good reason other than the code should probably not exist inside
of the class, this is in a helper function which, ironically, returns
the exact value the thing that calls it needs, allowing you to ponder
my utter insanity at splitting it apart.