Give the calendar the ability to mark days
This is simple change after the previous (not counting CSS, which has
gone back to not doing what a person probably wants) and allows the
config for the calendar component to take a list of days which should
be marked.

If this was used in a modal or a sidebar panel, it could target any
date and even have controls, and be able to mark what days are release
days on a video schedule, or holidays, or etc.