Update the calendar to allow more marks/day
This is a small change, but it allows for multiple indicators to be on
the same day, though at the moment we have no way of specifying that.

Observably this works, but we need more CSS mojo so that the indicator
item shares the space with the day rather than overlaying it if there
are too many marks.

It also does not properly wrap, but that's neither here nor there at
this point.