Allow for multiple dots with wrapping
A bit of searching has outlined that we can actually make the flex
container wrap (I am a dope), so that solves most problems. However the
example I found for this showed that even with a size for wrapping, if
a font is ever weirdly wide, the dots might still overlay, since it's
an em based size. So, we have wrapped the day number in its own element
so that we can z-index it above the dots just to make sure that it will
appear on top of them if anything untoward happens.