Ensure that we don't leak listeners like a boob
Upon close reflection (read: I had some debug logs that triggered more
than I thought they should) I realized that the code that was put in
place to be able to monitor when settings are changing so that
dependent settings were updated was actually leaking listeners.

Every time the display() function is invoked, it creates a new listener,
and since it is called every time the settings page opens, eventually
this becomes a Bad Deal (tm).

This change modifies the contract on the settings manager so that it
returns a cleanup function to remove the listener, which is then used
by the settings factory to create a cleanup function which can be
invoked by the settings page to clean things up.