Small tweak to our data reactive state
I would like to play around with allowing components to access the
settings portion of the data structure, but in the previous iteration
of the code, the 'this.settings' was set as an alias to the item from
within the data element BEFORE  it was made reactive.

Since the change code adjusts the settings object itself from a version
that is not reactive, the changes to settings don't trigger listeners
that care about it, which means that components would be able to see
the settings but only their initial version.

This is fixed by first loading the data and making a reactive proxy, and
then making our aliases from there.

This has to happen in two places since the data file can be edited
outside of Obsidian.