Year and month are bindable I guess?
I spent some time looking into those weird errors about capturing the
initial value and such.

As near as I can tell, you should never try to modify a prop because
they are supposed to be considered read-only, based on the theory that
if you DID modify them, you might think that the parent can see it, and
it can't.

However if you try to wrap a $state() around it, it gets mad because it
only captures the incoming initial value, which means that if the
parent changes the prop, you don't get to see like you normally would.

So it would appear you should use $bindable() for this because then it
declares the intent that what you change, the parent can see, and verse
visa, without things getting all janked.

I gather that when I'm doing here (or was) is valid and would only be a
problem if some other component was embedding this one, but doing it
this way ensures that things will work later if I did that, so its kind
of defensive.